Decouple signal handling from install process
